backend/requirements.txt
amqp==5.1.1
asgiref==3.5.1
async-timeout==4.0.2
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==38.0.3
Django==4.0.4
django-cors-headers==3.11.0
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
ecdsa==0.18.0
idna==3.3
kombu==5.2.4
mailjet-rest==1.3.4
nulltype==2.3.1
packaging==21.3
Pillow==9.1.0
plaid-python==7.4.0
prompt-toolkit==3.0.33
pyasn1==0.4.8
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
python-jose==3.3.0
pytz==2022.1
redis==4.3.5
requests==2.28.1
rsa==4.9
sib-api-v3-sdk==7.4.0
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
